requirements.txt
certifi==2018.8.24
opencv-contrib-python==3.4.3.18
pandas
torch>=0.4.1
numpy>=1.14.5
matplotlib==2.1.2
tqdm==4.23.4
scipy>=1.0.0
bcolz==1.2.1
Jinja2>=2.10.1
requests>=2.20.0
flask>=1.0.0
easydict==1.7
opencv_python>=3.4.0.12
Pillow==5.2.0
scikit_learn>=0.19.2
tensorboardX==1.2
torchvision==0.2.1
torchsummary
cython
imutils
zmq
